Auth refactoring
Replace the legacy auth user type with the one from the DB (with an added Permissions) field, to remove dependency on legacy auth for when we switch over.

Additionally, reshuffle the Auth folder a bit: separate out "who is currently signed in" into a Provider, and "who are they and what are their permissions" into a common layer. Finally, tidy up some imports.